:meat_on_bone: Much work done to layouts and content display:
added some logic and front matter so roles and credits etc are displayed on a given portfolio page
modified lots of typography so it looks nice and establishes a heirarchy. Not quiet there yet, but its in good shape
many of the templates were mofified in some way or another.